refactor(engine): extract InitResolver methods to their own TU
Moves Engine::init_apply_debug_raw_overrides_,
Engine::init_resolve_kv_dtype_policy_,
Engine::init_resolve_ssm_dtype_,
Engine::init_resolve_fp8_prefill_,
Engine::init_resolve_quant_flags_, and
Engine::init_compute_max_seq_len_ (~320 LOC across 6 methods) to
src/runtime/engine_init_resolver.cpp.

Declarations in engine.h unchanged. Bodies byte-identical.

Phase 4 of docs/superpowers/specs/2026-05-20-architecture-refactor-roadmap-design.md
