+++ CMakeLists.txt
@@ -220,6 +220,7 @@ set(IMP_RUNTIME_SOURCES
     src/runtime/request.cpp
     src/runtime/batch.cpp
     src/runtime/engine.cpp
+    src/runtime/engine_init_resolver.cpp
     src/runtime/engine_graph_decode.cpp
     src/runtime/vision_pipeline.cpp
     src/runtime/constraint_manager.cpp
